NetworkInteracting moved to package
