+++ Sources/Auth/Services/App/AppPairService.swift
@@ -1,6 +1,7 @@
 import Foundation
 import WalletConnectKMS
 import WalletConnectPairing
+import WalletConnectNetworking
 
 actor AppPairService {
     private let networkingInteractor: NetworkInteracting
