+++ Sources/Auth/Services/App/AppRespondSubscriber.swift
@@ -1,7 +1,8 @@
-import Combine
 import Foundation
-import WalletConnectUtils
+import Combine
 import JSONRPC
+import WalletConnectNetworking
+import WalletConnectUtils
 import WalletConnectPairing
 
 class AppRespondSubscriber {
