--- Sources/Auth/Types/RequestSubscriptionPayload.swift
@@ -1,7 +1,0 @@
-import Foundation
-import JSONRPC
-
-struct RequestSubscriptionPayload: Codable, Equatable {
-    let topic: String
-    let request: RPCRequest
-}
