--- Sources/Auth/Types/ResponseSubscriptionPayload.swift
@@ -1,7 +1,0 @@
-import Foundation
-import JSONRPC
-
-struct ResponseSubscriptionPayload: Codable, Equatable {
-    let topic: String
-    let response: RPCResponse
-}
